Issue #4575: Modified CyclomaticComplexityCheckTest.java and moved its input files to the cyclomaticcomplexity subdirectory
