+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/metrics/CyclomaticComplexityCheckTest.java
@@ -37,7 +37,7 @@ public class CyclomaticComplexityCheckTest
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "metrics" + File.separator + filename);
+                + "metrics" + File.separator + "cyclomaticcomplexity" + File.separator + filename);
     }
 
     @Test
@@ -51,7 +51,7 @@ public class CyclomaticComplexityCheckTest
             "4:5: " + getCheckMessage(MSG_KEY, 2, 0),
         };
 
-        verify(checkConfig, getPath("InputComplexitySwitchBlocks.java"), expected);
+        verify(checkConfig, getPath("InputCyclomaticComplexitySwitchBlocks.java"), expected);
     }
 
     @Test
@@ -65,7 +65,7 @@ public class CyclomaticComplexityCheckTest
             "4:5: " + getCheckMessage(MSG_KEY, 5, 0),
         };
 
-        verify(checkConfig, getPath("InputComplexitySwitchBlocks.java"), expected);
+        verify(checkConfig, getPath("InputCyclomaticComplexitySwitchBlocks.java"), expected);
     }
 
     @Test
@@ -88,7 +88,7 @@ public class CyclomaticComplexityCheckTest
             "104:13: " + getCheckMessage(MSG_KEY, 2, 0),
         };
 
-        verify(checkConfig, getPath("InputComplexity.java"), expected);
+        verify(checkConfig, getPath("InputCyclomaticComplexity.java"), expected);
     }
 
     @Test
@@ -122,6 +122,6 @@ public class CyclomaticComplexityCheckTest
         checkConfig.addAttribute("max", "100");
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
 
-        verify(checkConfig, getPath("InputComplexitySwitchBlocks.java"), expected);
+        verify(checkConfig, getPath("InputCyclomaticComplexitySwitchBlocks.java"), expected);
     }
 }
